Move BackToTop component to folder
